- Replaced Java serialization with streamable in NAKACK2, UNICAST3,   FlowControl, Locking, ExtendedUUID/SiteUUID  (https://issues.jboss.org/browse/JGRP-2033) - Moved RpcDispatcher$Marshaller into its own interface - Marshaller now (de-)serializes only RPC args and return values (incl exceptions)

+++ src/org/jgroups/protocols/DISCARD_PAYLOAD.java
@@ -26,7 +26,7 @@ public class DISCARD_PAYLOAD extends Protocol {
             Message msg=(Message)evt.getArg();
             if(msg.getLength() > 0) {
                 try {
-                    Long payload=(Long)msg.getObject();
+                    Long payload=msg.getObject();
                     if(payload != null) {
                         if(payload == seqno) {
                             synchronized(this) {
